requirements.txt
  
boto3==1.14.28
Django==3.1
redis==3.5.3
django-import-export==2.3.0
celery==4.4.7
fcm-django==0.3.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
gunicorn
django-cors-headers
psycopg2-binary
whitenoise
channels